[refactor] Preference 만드는 로직 분리
